+++ resources/views/components/header.blade.php
@@ -1,0 +1,8 @@
+<header class="container-fluid">
+    <div class="row">
+        <div class="col-xl-12">
+            <h1>{{$leftIcon}} {{$title}} {{$rightIcon}}</h1>
+            <hr class="h1-hr">
+        </div>
+    </div>
+</header>