+++ routes/web.php
@@ -3,3 +3,7 @@
 Route::get('/', function () {
     return view('home');
 });
+
+Route::get('/new-borrowing', function () {
+    return view('new-borrowing');
+});